spi-nor: Add support for n25q256a11 SPI flash device
We have a system that uses the n25q256a11 SPI flash device which is not
marked as supported in the 3.16 kernel series.

This patch is *not* for use in unstable, only Jessie.  A version between 3.16
and 4.7 (stretch) revamped how the chip support is determined, added explicit
support for this chip, and also added a way to add new chips at runtime
without requiring a recompile - all things that make this patch pointless on
unstable/testing.