ODROID-COMMON: mmc/block: add to validate MPT partition table on eMMC
The 'MPT' partition table is non-standard type which contains the partitions
at certain blocks on eMMC device to provide many partitions more than MBR
partition. When this partition table is discovered, standard partitions
such as MBR are ignored such that only the partitions from MPT are valid.

But there is a kernel crash bug when 'MPT' partition table is not exist.
This patch will check if 'MPT' partition is exist while checking its magic
number and proceed further more if it's valid.

Since Android needs to use 'MPT' partition, this cannot be simply eliminated.

Change-Id: Ic27fef3e28eab009affbb0f11a12472d58a373f6
Signed-off-by: Dongjin Kim <tobetter@gmail.com>