scsi: sg: Fix various bugs
Two fixes made between 3.16 and 4.4:
* sg: O_EXCL and other lock handling
* sg: prevent integer overflow when converting from sectors to bytes

A 3.16-specific patch to make the rest apply cleanly:
* scsi: sg: Change next_cmd_len handling to mirror upstream

Then a large number of fixes that have been applied to 4.4-stable but
not yet to 3.16-stable.  Only the last of these has a CVE ID, but
several others do fix bugs with some security impact.