Build-Depend on libbabeltrace-ctf-dev with no alternative
buildds for stable suites have a stupid configuration that only
considers the first of each alternation in Build-Depends.  So for
the dependency on:

    libbabeltrace-dev (>= 1.5.3-2~) | libbabeltrace-ctf-dev (>= 1.5.0)

they check that libbabeltrace-dev is available, find that its version
is too old, and then give up.

For this backport, Build-Depend only on libbabeltrace-ctf-dev, which
is what stretch has.