ODROID-N2: arm64/dts: Move gpio-line-names into the periphs_pinctrl node
In general, the gpio-line-names property should belong to the node
whereby the gpio-controller property by the related document written.
But from now, at least for the Odroid boards based on AMLogic SoC,
the property has to be in the periphs_pinctrl node so that the
function for assigning the line names recognizes that.

Signed-off-by: Deokgyu Yang <secugyu@gmail.com>
Change-Id: I1b491db19dd7e6b678d34c4d288de7549458d783