arm64: dts: meson-sm1: add uart_<A|AO_B|B|C> nodes
ODROID-C4 can export upto 3 UART ports on J2 (40 pin header) and 1 UART
port on J7 (7 pin header). This patch is to add their UART nodes as
disabled status since the pins are dedicated for GPIO. The pins can be
switched to UART when the status is changed to "okay" by editing the
device tree file or using device tree overlay.

Signed-off-by: Dongjin Kim <tobetter@gmail.com>
Change-Id: Icad96ae644e00aedbdfedcd86fcb2e964c1abff5